updates based on codereview
